Relocate current submission error functionality to new Service and Controller.
